backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
django-cors-headers==4.4.0
django-filter==24.3
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
celery==5.4.0
redis==5.0.8
Pillow==10.4.0
drf-spectacular==0.27.2
twilio==9.2.0
psycopg2-binary==2.9.9
scikit-learn==1.5.1
nltk==3.9.1
